fix: killProcessOnPort and wait idle
