+++ packages/typescript-private/device-server/src/internal/util/net.ts
@@ -1,6 +1,5 @@
-import { delay, loop } from '@dogu-tech/common';
+import { loop } from '@dogu-tech/common';
 import { findFreePorts } from 'find-free-ports';
-import net from 'net';
 
 const startPort = 20000;
 const endPort = 60000;
@@ -35,35 +34,5 @@ function cleanUpUsedPorts(): void {
     if (now - accessTime > accessBlockTime) {
       usedportToAccessTime.delete(usedPort);
     }
-  });
-}
-
-export async function waitPortOpen(port: number, timeout = 10000): Promise<void> {
-  const start = Date.now();
-  while (Date.now() - start < timeout) {
-    const isFree = await isFreePort(port).catch(() => true);
-    if (!isFree) {
-      return;
-    }
-    await delay(300);
-  }
-  throw Error(`waitPortOpen. port:${port} is not open`);
-}
-
-export function isFreePort(port: number): Promise<boolean> {
-  return new Promise((accept, reject) => {
-    const sock = net.connect({ host: '127.0.0.1', port: port });
-    sock.once('connect', () => {
-      sock.end();
-      accept(false);
-    });
-    sock.once('error', (e: NodeJS.ErrnoException) => {
-      sock.destroy();
-      if (e.code === 'ECONNREFUSED') {
-        accept(true);
-      } else {
-        reject();
-      }
-    });
   });
 }
